ghostline/api/pyproject.toml
[tool.poetry]
name = "ghostline-api"
version = "0.1.0"
description = "GhostLine API Service"
authors = ["GhostLine Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.14"}
uvicorn = "^0.34.3"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.10"
redis = "^5.2.0"
celery = "^5.4.0"
pydantic-settings = "^2.7.0"
boto3 = "^1.35.0"
httpx = "^0.28.1"
alembic = "^1.14.0"
pgvector = "^0.3.0"
asyncpg = "^0.30.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
requests = "^2.32.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^0.25.0"
pytest-cov = "^6.0.0"
black = "^24.0.0"
ruff = "^0.9.0"
mypy = "^1.14.0"
types-redis = "^4.6.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = [
    ".git",
    "__pycache__",
    "docs",
    "build",
    "dist",
    ".venv",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "B008",  # Do not perform function call in argument defaults (needed for FastAPI Depends)
    "E712",  # Comparison to True (sometimes needed for SQLAlchemy filters)
]

[tool.ruff.lint.per-file-ignores]
"scripts/*" = ["E402", "F403", "E501"]  # Allow late imports, star imports, and long lines in scripts
"alembic/*" = ["F403"]  # Allow star imports in Alembic migrations
"app/models/book_outline.py" = ["E501"]  # Allow long comment line

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
strict_optional = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/migrations/*"]

